Update to Readme
Change 1 Original readme was a .md file to that most people who this repository is made for do not know what .md is for so changed to .txt
Change 2 wrote in the beginning text to the readme for people to add onto to help us define what the project is to be. The discussion on discord was a mess of what kind of game. I am hoping they can decide on 1 then put it into here. Followed by a list of features then how to use them.